--- letchat/pages/jobboard2.tsx
@@ -1,57 +1,0 @@
-import Head from "next/head";
-import { Inter } from "@next/font/google";
-import Ticket from "../components/Ticket";
-import Content from "../components/Content";
-import NavigationPoint from "../components/NavigationPoint";
-
-import { getAuth } from "firebase/auth";
-import { useAuthState } from "react-firebase-hooks/auth";
-import { useRouter } from "next/router";
-import { initFirebase } from "../firebase/firebaseApp";
-import Link from "next/link";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export default function JobBoard() {
-  initFirebase();
-  const auth = getAuth();
-  const [user, loading] = useAuthState(auth);
-  const router = useRouter();
-  /*   if (!user) {
-    router.push("/login");
-    return <div>Please sign in to continue</div>;
-  } */
-
-  function openTicket(id: number) {
-    console.log(id);
-  }
-  if (user) {
-    return (
-      <>
-        <Head>
-          <title>LetChat | Job Board</title>
-          <meta name="description" content="Generated by create next app" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <main>
-          <button onClick={() => auth.signOut()}>Firebase LogOut</button>
-
-          <Content
-            title="Messages"
-            navigation={<NavigationPoint />}
-            central={<Ticket openTicket={openTicket} />}
-          />
-        </main>
-      </>
-    );
-  }
-  if (!user) {
-    return (
-      <>
-        <div>You're not logged in</div>
-        <button onClick={() => router.push("/login")}>Back</button>
-      </>
-    );
-  }
-}
